[A11y] Make the language selector accessible
By splitting the language selector into its own cell we can make it accessible.

Fixes BXC #53729
