+++ main/src/core/MonoDevelop.Ide/MonoDevelop.Ide.Projects/GtkTemplateCellRenderer.cs
@@ -38,19 +38,11 @@ namespace MonoDevelop.Ide.Projects
 {
 	class GtkTemplateCellRenderer : CellRendererText
 	{
-		Rectangle languageRect;
-		int dropdownTriangleWidth = 8;
-		int dropdownTriangleHeight = 5;
-		const int dropdownTriangleRightHandPadding = 8;
-		const int languageRightHandPadding = 4;
-		const int languageLeftHandPadding = 9;
 		const int iconTextPadding = 9;
 		int groupTemplateHeadingTotalYPadding = 24;
 		int recentTemplateHeadingTotalYPadding = 30;
 		const int groupTemplateHeadingYOffset = 4;
 		const int categoryTextPaddingX = 4;
-
-		int minLanguageRectWidth;
 
 		public SolutionTemplate Template { get; set; }
 		public string SelectedLanguage { get; set; }
@@ -60,11 +52,6 @@ namespace MonoDevelop.Ide.Projects
 
 		public GtkTemplateCellRenderer ()
 		{
-			minLanguageRectWidth = languageLeftHandPadding +
-				dropdownTriangleWidth +
-				dropdownTriangleRightHandPadding +
-				languageRightHandPadding + 10;
-
 			if (IsYosemiteOrHigher ()) {
 				groupTemplateHeadingTotalYPadding -= 1;
 				recentTemplateHeadingTotalYPadding -= 1;
@@ -74,16 +61,6 @@ namespace MonoDevelop.Ide.Projects
 		static bool IsYosemiteOrHigher ()
 		{
 			return Platform.IsMac && (Platform.OSVersion >= MacSystemInformation.Yosemite);
-		}
-
-		public bool IsLanguageButtonPressed (EventButton button)
-		{
-			return !RenderRecentTemplate && languageRect.Contains ((int)button.X, (int)button.Y);
-		}
-
-		public Rectangle GetLanguageRect ()
-		{
-			return languageRect;
 		}
 
 		public override void GetSize (Widget widget, ref Rectangle cell_area, out int x_offset, out int y_offset, out int width, out int height)
@@ -108,45 +85,9 @@ namespace MonoDevelop.Ide.Projects
 
 					Rectangle iconRect = DrawIcon (ctx, widget, cell_area, flags);
 
-					if (!RenderRecentTemplate && (!Template.AvailableLanguages.Any () || !IsTemplateRowSelected (widget, flags))) {
-						DrawTemplateNameText (window, widget, cell_area, iconRect, Rectangle.Zero, flags);
-						return;
-					}
-
-					int textHeight = 0;
-					int textWidth = 0;
-
-					SetMarkup (layout, GetSelectedLanguage ());
-					layout.GetPixelSize (out textWidth, out textHeight);
-
-					languageRect = GetLanguageButtonRectangle (window, widget, cell_area, textHeight, textWidth);
-
-					DrawTemplateNameText (window, widget, cell_area, iconRect, languageRect, flags);
+					DrawTemplateNameText (window, widget, cell_area, iconRect, flags);
 					if (RenderRecentTemplate)
-						DrawCategoryText (ctx, widget, cell_area, iconRect, languageRect, flags);
-
-
-					StateType state = StateType.Normal;
-					if (!RenderRecentTemplate) {
-						RoundBorder (ctx, languageRect.X, languageRect.Y, languageRect.Width, languageRect.Height);
-						SetSourceColor (ctx, Styles.NewProjectDialog.TemplateLanguageButtonBackground.ToCairoColor ());
-						ctx.Fill ();
-					} else
-						state = GetState (widget, flags);
-
-					int languageTextX = languageRect.X + languageLeftHandPadding;
-					if (!TemplateHasMultipleLanguages ()) {
-						languageTextX = languageRect.X + (languageRect.Width - textWidth) / 2;
-					}
-					int languageTextY = languageRect.Y + (languageRect.Height - textHeight) / 2;
-
-					window.DrawLayout (widget.Style.TextGC (state), languageTextX, languageTextY, layout);
-
-					if (TemplateHasMultipleLanguages ()) {
-						int triangleX = languageTextX + textWidth + languageRightHandPadding;
-						int triangleY = languageRect.Y + (languageRect.Height - dropdownTriangleHeight) / 2;
-						DrawTriangle (ctx, triangleX, triangleY);
-					}
+						DrawCategoryText (ctx, widget, cell_area, iconRect, flags);
 				}
 			}
 		}
@@ -184,14 +125,14 @@ namespace MonoDevelop.Ide.Projects
 			return iconRect;
 		}
 
-		void DrawTemplateNameText (Drawable window, Widget widget, Rectangle cell_area, Rectangle iconRect, Rectangle languageRect, CellRendererState flags)
+		void DrawTemplateNameText (Drawable window, Widget widget, Rectangle cell_area, Rectangle iconRect, CellRendererState flags)
 		{
 			StateType state = GetState (widget, flags);
 
 			using (var layout = new Pango.Layout (widget.PangoContext)) {
 
 				layout.Ellipsize = Pango.EllipsizeMode.End;
-				int textPixelWidth = widget.Allocation.Width - ((int)Xpad * 2) - iconRect.Width - iconTextPadding - languageRect.Width;
+				int textPixelWidth = cell_area.Width - ((int)Xpad * 2) - iconRect.Width - iconTextPadding;
 				layout.Width = (int)(textPixelWidth * Pango.Scale.PangoScale);
 
 				layout.SetMarkup (GLib.Markup.EscapeText (Template.Name));
@@ -204,7 +145,7 @@ namespace MonoDevelop.Ide.Projects
 			}
 		}
 
-		void DrawCategoryText (Cairo.Context ctx, Widget widget, Rectangle cell_area, Rectangle iconRect, Rectangle languageRect, CellRendererState flags)
+		void DrawCategoryText (Cairo.Context ctx, Widget widget, Rectangle cell_area, Rectangle iconRect, CellRendererState flags)
 		{
 			StateType state = GetState (widget, flags);
 			var isSelected = state == StateType.Selected || state == StateType.Active;
@@ -212,7 +153,7 @@ namespace MonoDevelop.Ide.Projects
 			using (var layout = new Pango.Layout (widget.PangoContext)) {
 
 				layout.Ellipsize = Pango.EllipsizeMode.End;
-				int textPixelWidth = widget.Allocation.Width - ((int)Xpad * 2) - iconRect.Width - iconTextPadding - languageRect.Width;
+				int textPixelWidth = cell_area.Width - ((int)Xpad * 2) - iconRect.Width - iconTextPadding;
 				layout.Width = (int)(textPixelWidth * Pango.Scale.PangoScale);
 				layout.FontDescription = Fonts.FontExtensions.CopyModified (widget.Style.FontDesc, -1);
 
@@ -228,12 +169,6 @@ namespace MonoDevelop.Ide.Projects
 			}
 		}
 
-		static bool IsTemplateRowSelected (Widget widget, CellRendererState flags)
-		{
-			StateType stateType = GetState (widget, flags);
-			return (stateType == StateType.Selected) || (stateType == StateType.Active);
-		}
-
 		static StateType GetState (Widget widget, CellRendererState flags)
 		{
 			StateType stateType = StateType.Normal;
@@ -246,80 +181,6 @@ namespace MonoDevelop.Ide.Projects
 			if ((flags & CellRendererState.Selected) != 0)
 				stateType = widget.HasFocus ? StateType.Selected : StateType.Active;
 			return stateType;
-		}
-
-		string GetSelectedLanguage ()
-		{
-			if (!Template.AvailableLanguages.Any ())
-				return String.Empty;
-			else if (RenderRecentTemplate)
-				return Template.Language;
-			else if (Template.AvailableLanguages.Contains (SelectedLanguage))
-				return SelectedLanguage;
-
-			return Template.AvailableLanguages.First ();
-		}
-
-		void SetMarkup (Pango.Layout layout, string text)
-		{
-			string markup = "<span size='smaller'>" + text + "</span>";
-			layout.SetMarkup (markup);
-		}
-
-		Rectangle GetLanguageButtonRectangle (Drawable window, Widget widget, Rectangle cell_area, int textHeight, int textWidth)
-		{
-			int languageRectangleHeight = cell_area.Height - 8;
-			int languageRectangleWidth = textWidth + languageLeftHandPadding;
-			if (TemplateHasMultipleLanguages ()) {
-				languageRectangleWidth += languageRightHandPadding + dropdownTriangleWidth + dropdownTriangleRightHandPadding;
-			} else {
-				languageRectangleWidth += languageLeftHandPadding;
-				languageRectangleWidth = Math.Max (languageRectangleWidth, minLanguageRectWidth);
-			}
-
-			var dy = (cell_area.Height - languageRectangleHeight) / 2 - 1;
-			var y = cell_area.Y + dy;
-			var x = widget.Allocation.Width - languageRectangleWidth - (int)Xpad;
-
-			return new Rectangle (x, y, languageRectangleWidth, languageRectangleHeight);
-		}
-
-		bool TemplateHasMultipleLanguages ()
-		{
-			return !RenderRecentTemplate && Template.AvailableLanguages.Count > 1;
-		}
-
-		void DrawTriangle (Cairo.Context ctx, int x, int y)
-		{
-			int width = dropdownTriangleWidth;
-			int height = dropdownTriangleHeight;
-
-			SetSourceColor (ctx, Styles.NewProjectDialog.TemplateLanguageButtonTriangle.ToCairoColor ());
-			ctx.MoveTo (x, y);
-			ctx.LineTo (x + width, y);
-			ctx.LineTo (x + (width / 2), y + height);
-			ctx.LineTo (x, y);
-			ctx.Fill ();
-		}
-
-		// Taken from MonoDevelop.Components.SearchEntry.
-		static void RoundBorder (Cairo.Context ctx, double x, double y, double w, double h)
-		{
-			double r = h / 2;
-			ctx.Arc (x + r, y + r, r, Math.PI / 2, Math.PI + Math.PI / 2);
-			ctx.LineTo (x + w - r, y);
-
-			ctx.Arc (x + w - r, y + r, r, Math.PI + Math.PI / 2, Math.PI + Math.PI + Math.PI / 2);
-
-			ctx.LineTo (x + r, y + h);
-
-			ctx.ClosePath ();
-		}
-
-		// Taken from Mono.TextEditor.HelperMethods.
-		public static void SetSourceColor (Cairo.Context cr, Cairo.Color color)
-		{
-			cr.SetSourceRGBA (color.R, color.G, color.B, color.A);
 		}
 	}
 }
